+++ internal/processor/filterset/config_test.go
@@ -21,13 +21,13 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/config/configparser"
 	"go.opentelemetry.io/collector/internal/processor/filterset/regexp"
 )
 
 func readTestdataConfigYamls(t *testing.T, filename string) map[string]*Config {
 	testFile := path.Join(".", "testdata", filename)
-	v, err := config.NewParserFromFile(testFile)
+	v, err := configparser.NewParserFromFile(testFile)
 	require.NoError(t, err)
 
 	cfgs := map[string]*Config{}